+++ module/const.js
@@ -2,8 +2,10 @@
  * Analytics.
  *
  *      ./module/const.js
- *      v0.0.1
+ *      v0.0.2
  */
+
+export const DEBUG = false;
 
 /**
 * console colors.
@@ -19,9 +21,10 @@ export const CONSOLE_DEFAULT = CONSOLE_GREY;
 * modules and versions.
 */
 
-export const MIN_FOUNDRY_VERSION = "9.0";
-export const CM_MODULE_NAME      = "analytics";
-export const CM_NAME             = "Analytics";
-export const CM_LABEL            = "*** | ";
-export const CM_VERSION          = "0.0.1";
-export const CM_ENABLED          = "enabled";
+export const MIN_FOUNDRY_VERSION = "9.242";
+export const MIN_DND5E_VERSION   = "1.5.7";
+export const MODULE_NAME      	 = "analytics";
+export const NAME             	 = "Analytics";
+export const LABEL            	 = "*** | ";
+export const VERSION          	 = "0.0.1";
+export const ENABLED          	 = "enabled";