+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v0.0.2 (2022-1-23)
+
+* Actors and Items Dialog. D&D5e dependency.
+
 # v0.0.1 (2022-1-17)
 
-* Initial release.
+* Initial release. Basic module functionality with Toolbar and Dialogs.